add logging function and update exception when user push ctrl+c

+++ solution_installer.py
@@ -6,6 +6,14 @@ from xml.etree.ElementTree import parse
 import shutil
 import shlex
 from os import environ
+import signal
+
+
+def logdebug(msg):
+    print '\033[92m' + msg.replace('\n', '') + '\x1b[0m' 
+
+def loginfo(msg):
+    print msg.replace('\n', '')
 
 
 def find_solution(taget_local_path, target_solution):
@@ -17,10 +25,9 @@ def find_solution(taget_local_path, target_solution):
                 return root
 
 
-def get_src_from_gitbase(git_path, taget_local_path):
+def get_src_from_gitbase(solution_repo_git_path, taget_local_path):
     if not os.path.isdir(taget_local_path):
-        print 'git clone from %s: ' % git_path
-        cmd = 'git clone %s %s' % (git_path, taget_local_path)
+        cmd = 'git clone %s %s' % (solution_repo_git_path, taget_local_path)
         run_command(cmd)
     else:
         os.chdir(taget_local_path)
@@ -44,28 +51,29 @@ def launch_target_solution(ws_path, target_solution, concert_name):
     run_command(cmd, multiple_cmd=True, executable='/bin/bash')
 
 
-def run_command(command, multiple_cmd=False, executable='/bin/sh'):
+def run_command(command, multiple_cmd=False, executable=None):
+    logdebug(command)
     if multiple_cmd:
         process = subprocess.Popen(command, shell=multiple_cmd, executable=executable, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     else:
-        process = subprocess.Popen(shlex.split(command), stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-
+        process = subprocess.Popen(shlex.split(command), executable=executable, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
 
     while True:
         output = process.stdout.readline()
         if output == '' and process.poll() is not None:
             break
         if output:
-            print output.strip()
-    print process.stderr.readline()
+            loginfo(output.strip())
+    loginfo("Return Code of Process: [%s]" % process.returncode)
+    for err in process.stderr:
+        loginfo(err)
+    
 
 
 def get_ros_ws():
     cmake_prefix_path = os.getenv('CMAKE_PREFIX_PATH')
     workspace_path = ''
-    if cmake_prefix_path is None or len(cmake_prefix_path) is 0:
-        print 'Not set workspace'
-    else:
+    if cmake_prefix_path is not None and len(cmake_prefix_path) is not 0:
         for ws in cmake_prefix_path.split(':'):
             if os.path.isfile(os.path.join(ws, '.catkin')) and os.path.isfile(os.path.join(ws[:ws.rindex('/')], '.catkin_workspace')):
                 workspace_path = ws[:ws.rindex('/')]
@@ -76,29 +84,34 @@ if __name__ == '__main__':
 
     target_solution = 'test_concert'
     concert_name = 'test.concert'
+    solution_repo_name = 'dwlee/concert_common_solutions'
 
-    repo_name = 'dwlee/concert_common_solutions'
     taget_local_path = os.path.join(tempfile.gettempdir(), 'solution_repo')
-    git_path = 'https://github.com/%s.git' % (repo_name)
-
-    ros_ws_path = get_ros_ws()
-    # get solution from repo
-    get_src_from_gitbase(git_path, taget_local_path)
-    ros_ws_src_path = os.path.join(ros_ws_path, 'src')
-    if not os.path.isdir(ros_ws_src_path):
-        print 'Do not set ROS workspace: [%s]' % ros_ws_src_path
-    else:
-        src = find_solution(taget_local_path, target_solution)
-        dst = os.path.join(ros_ws_src_path, target_solution)
-        if not os.path.isdir(dst):
-            print 'create solution in workpace: [%s]' % dst
+    solution_repo_git_path = 'https://github.com/%s.git' % (solution_repo_name)
+    try:
+        ros_ws_path = get_ros_ws()
+        if len(ros_ws_path):
+            # get solution from repo
+            get_src_from_gitbase(solution_repo_git_path, taget_local_path)
+            ros_ws_src_path = os.path.join(ros_ws_path, 'src')
+            if not os.path.isdir(ros_ws_src_path):
+                loginfo('Do not set ROS workspace: [%s]' % ros_ws_src_path)
+            else:
+                src = find_solution(taget_local_path, target_solution)
+                dst = os.path.join(ros_ws_src_path, target_solution)
+                if not os.path.isdir(dst):
+                    loginfo('Create solution in workpace: [%s]' % dst)
+                else:
+                    loginfo('Already existed solution in workpace: [%s]' % dst)
+                    shutil.rmtree(dst)
+                shutil.copytree(src, dst)
+            # build solution
+            check_dependency(ros_ws_path)
+            build_workspace(ros_ws_path)
+            # launch target solution
+            launch_target_solution(ros_ws_path, target_solution, concert_name)
         else:
-            print 'already existed solution in workpace: [%s]' % dst
-            shutil.rmtree(dst)
-        shutil.copytree(src, dst)
-    # build solution
-    check_dependency(ros_ws_path)
-    build_workspace(ros_ws_path)
-    # launch target solution
-    launch_target_solution(ros_ws_path, target_solution, concert_name)
-    print 'Bye Bye'
+            loginfo('Do not set ROS workspace: [%s]' % ros_ws_path)
+    except KeyboardInterrupt as e:
+        loginfo('Stop all process: [%s]' % str(e))
+    loginfo('Bye Bye')
